Rename vguest to virtio_net
Exec selftest as:

sudo SNABB_TEST_VIRTIO_PCIDEV=0000:00:03.0 \
     ./snabb snsh -t apps.virtio_net.virtio_net_app
