+++ src/apps/virtio_net/virtio_net_app.lua
@@ -2,22 +2,22 @@ module(..., package.seeall)
 
 local basic_apps = require("apps.basic.basic_apps")
 local pcap       = require("apps.pcap.pcap")
-local virtio_dev = require("apps.vguest.vguest_dev")
+local virtio_dev = require("apps.virtio_net.virtio_net_dev")
 
-local VGuest = {}
-VGuest.__index = VGuest
+local virtio_net = {}
+virtio_net.__index = virtio_net
 
-function VGuest:new(args)
+function virtio_net:new(args)
    return setmetatable({
       device = assert(virtio_dev.VGdev:new(args)),
    }, self)
 end
 
-function VGuest:stop()
+function virtio_net:stop()
    self.device:close()
 end
 
-function VGuest:push()
+function virtio_net:push()
    local dev = self.device
    local l = self.input.rx
    if not dev or not l then return end
@@ -28,7 +28,7 @@ function VGuest:push()
    dev:sync_transmit()
 end
 
-function VGuest:pull()
+function virtio_net:pull()
    local dev = self.device
    local l = self.output.tx
    if not dev or not l then return end
@@ -40,7 +40,7 @@ function VGuest:pull()
    self:add_receive_buffers()
 end
 
-function VGuest:add_receive_buffers()
+function virtio_net:add_receive_buffers()
    local dev = self.device
    while dev:can_add_receive_buffer() do
       dev:add_receive_buffer(packet.allocate())
@@ -56,10 +56,10 @@ function selftest()
    engine.configure(config.new())
    local c = config.new()
    config.app(c, 'source', pcap.PcapReader, input_file)
-   config.app(c, 'vguest', VGuest, {pciaddr=pcidev})
+   config.app(c, 'virtio_net', virtio_net, {pciaddr=pcidev})
    config.app(c, 'sink', basic_apps.Sink)
-   config.link(c, 'source.output -> vguest.rx')
-   config.link(c, 'vguest.tx -> sink.input')
+   config.link(c, 'source.output -> virtio_net.rx')
+   config.link(c, 'virtio_net.tx -> sink.input')
    engine.configure(c)
    engine.main({duration = 1, report={showlinks=true, showapps=true}})
 end
